+++ src/App.tsx
@@ -2,7 +2,7 @@ import React, { useContext } from 'react';
 import { Header } from './components/Header';
 import { Main } from './components/Main';
 import { Footer } from './components/Footer';
-import { StateContext } from './components/TodosContext';
+import { StateContext } from './context/TodosContext';
 
 export const App: React.FC = () => {
   const { todos } = useContext(StateContext);
