+++ src/types/FilterType.ts
@@ -1,0 +1,5 @@
+export enum FilterType {
+  ALL = 'all',
+  ACTIVE = 'active',
+  COMPLETED = 'completed',
+}
