+++ src/types/State.ts
@@ -1,0 +1,7 @@
+import { FilterType } from './FilterType';
+import { Todo } from './Todo';
+
+export interface State {
+  todos: Todo[],
+  filterBy: FilterType,
+}
