+++ src/types/Todo.ts
@@ -3,14 +3,3 @@ export interface Todo {
   title: string,
   completed: boolean,
 }
-
-export enum FilterType {
-  ALL = 'all',
-  ACTIVE = 'active',
-  COMPLITED = 'completed',
-}
-
-export interface State {
-  todos: Todo[],
-  filterBy: FilterType,
-}
